Cache HDF5 objects in UVH5File, pull out Vis
UVH5File can be used to get HDF5 objects from the underlying UVH5 file.
HDF5 objects obtained this way are cached so that future requests will
return the same HDF5 object.  Also, Vis is no longer a field of
UVH5File, but this may change back if it proves cumbersome.